change total molar mass calculation
similar to mass, the composite fraction is defined by absolute moles now. The way total molar mass was being calculated before did not take into account the relative amounts of each component and didn't make sense.

everywhere where something was specified as 0.5 mol fraction is now 1 mol

e.g. 1 mol NiCd  = 1 mol Ni + 1 mol Cd

So the molar masses of these compounds have changed but not the energies of electrochemical as this happened to work out with the previous method.

This is assuming a valence (n_e) of 1 per mol for electrochemical reactions. n_e should be explicitly set at some point in the future