add kwargs to join_col_vals
by default join_col_vals was making a set and sorting the values. This changes the order of different columns. this is now a kwarg and was not applied to the source column and specific price colmns in mat_data so that the source of specific prices can be traced